Factor push and create events into modules
