+++ src/models/events/payload.rs
@@ -1,4 +1,9 @@
+mod create;
+mod push;
+
 use crate::models::repos::GitUser;
+pub use create::*;
+pub use push::*;
 use reqwest::Url;
 use serde::{Deserialize, Serialize};
 
@@ -14,31 +19,6 @@ pub enum EventPayload {
     PushEvent(PushEventPayload),
     CreateEvent(CreateEventPayload),
     UnknownEvent(serde_json::Value),
-}
-
-/// The payload in a [`EventPayload::PushEvent`] type.
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
-#[non_exhaustive]
-pub struct PushEventPayload {
-    pub push_id: u64,
-    pub size: u64,
-    pub distinct_size: u64,
-    pub r#ref: String,
-    pub head: String,
-    pub before: String,
-    pub commits: Vec<Commit>,
-}
-
-/// The payload in a [`EventPayload::CreateEvent`] type.
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
-#[non_exhaustive]
-pub struct CreateEventPayload {
-    // a null ref will occur on the initial create event
-    pub r#ref: Option<String>,
-    pub ref_type: String,
-    pub master_branch: String,
-    pub description: Option<String>,
-    pub pusher_type: String,
 }
 
 /// A git commit in specific payload types.
